+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## [Unreleased][]
 
-[Unreleased]: https://github.com/chaostoolkit/chaostoolkit-lib/compare/1.27.1...HEAD
+[Unreleased]: https://github.com/chaostoolkit/chaostoolkit-lib/compare/1.27.2...HEAD
+
+## [1.27.2][] - 2022-03-23
+
+[1.27.2]: https://github.com/chaostoolkit/chaostoolkit-lib/compare/1.27.1...1.27.2
 
 ### Changed